+++ Chocolatey/tools/chocolateyinstall.ps1
@@ -4,14 +4,14 @@ $toolsDir   = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
 $packageArgs = @{
   packageName   = $env:ChocolateyPackageName
   fileType      = 'exe'
-  url           = 'https://github.com/vertopolkaLF/silence/releases/download/v1.4.1/silence-v1.4-x86-setup.exe'
-  url64bit      = 'https://github.com/vertopolkaLF/silence/releases/download/v1.4.1/silence-v1.4-x64-setup.exe'
+  url           = 'https://github.com/vertopolkaLF/silence/releases/download/v1.5/silence-v1.5-x86-setup.exe'
+  url64bit      = 'https://github.com/vertopolkaLF/silence/releases/download/v1.5/silence-v1.5-x64-setup.exe'
   
   softwareName  = 'silence!*'
   
-  checksum      = 'E241BD73443955A2338A37AFB63AF9DBC1515CEEB61A2A2C20648A7023DA1A00'
+  checksum      = 'PLACEHOLDER_X86_CHECKSUM'
   checksumType  = 'sha256'
-  checksum64    = '6D926E3C4AB5CC4F7A5F0988F673360A33654A24318349545EEFA184383EB2A5'
+  checksum64    = 'PLACEHOLDER_X64_CHECKSUM'
   checksumType64= 'sha256'
   
   silentArgs    = '/VERYSILENT /SUPPRESSMSGBOXES /NORESTART /SP-'